requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
Pillow>=10.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.0
django-cors-headers>=4.0
openpyxl>=3.1
qrcode[pil]>=7.4
